upgrade the java version to 21 and the Spring Boot version to 3.2.1. remove the dependency okhttp3 and fix the tests and documentations.
