+++ pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>3.1.3</version>
+        <version>3.2.1</version>
         <relativePath/>
     </parent>
     <groupId>com.github.ksewen</groupId>
@@ -15,14 +15,12 @@
     <description>Enterprise-level Spring Cloud integration project.</description>
     <properties>
         <revision>0.3.0</revision>
-        <java.version>17</java.version>
-        <spring-cloud.version>2022.0.4</spring-cloud.version>
+        <java.version>21</java.version>
+        <spring-cloud.version>2023.0.0</spring-cloud.version>
         <springdoc.version>2.2.0</springdoc.version>
         <yorozuya.version>${revision}</yorozuya.version>
         <guava.version>32.1.2-jre</guava.version>
-        <jjwt.version>0.12.3</jjwt.version>
-        <ok.http.3.version>4.11.0</ok.http.3.version>
-        <http.client.version>5.2.1</http.client.version>
+        <http.client.version>5.3</http.client.version>
         <swagger.annotations.jakarta.version>2.2.15</swagger.annotations.jakarta.version>
 
         <maven.jacoco.plugin.version>0.8.10</maven.jacoco.plugin.version>
@@ -90,13 +88,6 @@
                 <version>${guava.version}</version>
             </dependency>
 
-            <!-- jjwt -->
-            <dependency>
-                <groupId>io.jsonwebtoken</groupId>
-                <artifactId>jjwt</artifactId>
-                <version>${jjwt.version}</version>
-            </dependency>
-
             <!-- Springdoc -->
             <dependency>
                 <groupId>org.springdoc</groupId>
@@ -110,11 +101,6 @@
             </dependency>
 
             <!-- HttpClient -->
-            <dependency>
-                <groupId>com.squareup.okhttp3</groupId>
-                <artifactId>okhttp</artifactId>
-                <version>${ok.http.3.version}</version>
-            </dependency>
             <dependency>
                 <groupId>org.apache.httpcomponents.client5</groupId>
                 <artifactId>httpclient5</artifactId>
