+++ yorozuya-auth/yorozuya-auth-server/src/test/java/com/github/ksewen/yorozuya/auth/server/configuration/SecurityPropertiesTest.java
@@ -2,12 +2,16 @@ package com.github.ksewen.yorozuya.auth.server.configuration;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.*;
+import static org.mockito.Mockito.mock;
 
 import com.github.ksewen.yorozuya.auth.server.annotation.EnableAuthServer;
 import java.util.Map;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.servlet.handler.HandlerMappingIntrospector;
 
 /**
  * @author ksewen
@@ -18,7 +22,8 @@ import org.springframework.boot.test.context.SpringBootTest;
     classes = {
       SecurityAutoConfiguration.class,
       AuthServerAutoConfigurationTest.MockJwtToUserConverterAutoConfiguration.class,
-      SecurityAutoConfigurationTest.MockWebEndpointPropertiesAutoConfiguration.class
+      SecurityAutoConfigurationTest.MockWebEndpointPropertiesAutoConfiguration.class,
+      SecurityPropertiesTest.MockHandlerMappingIntrospectorAutoConfiguration.class
     },
     properties = {
       "security.permit-urls.get=/auth/**",
@@ -45,5 +50,14 @@ class SecurityPropertiesTest {
   void isProtectManagementEndpoints() {
     boolean result = this.securityProperties.isProtectManagementEndpoints();
     assertThat(result).isFalse();
+  }
+
+  @Configuration
+  static class MockHandlerMappingIntrospectorAutoConfiguration {
+
+    @Bean
+    public HandlerMappingIntrospector mvcHandlerMappingIntrospector() {
+      return mock(HandlerMappingIntrospector.class);
+    }
   }
 }
