+++ yorozuya-samples/auth-server/src/main/resources/application.yml
@@ -30,15 +30,16 @@ spring:
       client:
         config:
           default:
-            connect-timeout: ${common.ok.http.client.connect-timeout}
-            read-timeout: ${common.ok.http.client.read-timeout}
+            connectTimeout: ${common.http.client.hc5.connect-timeout}
+            read-timeout: ${common.http.client.hc5.socket-timeout}
           #            logger-level: FULL
           restClient:
             url: http://127.0.0.1:8080/
           invalidRestClient:
             url: http://127.0.0.1:38080/
-      okhttp:
-        enabled: true
+      httpclient:
+        hc5:
+          enabled: true
   messages:
     encoding: UTF-8
     basename: validation-messages
@@ -49,12 +50,11 @@ eureka:
     fetch-registry: false
 
 common:
-  ok:
-    http:
-      client:
+  http:
+    client:
+      hc5:
         connect-timeout: 3000
-        read-timeout: 5000
-        write-timeout: 5000
+        socket-timeout: 5000
 
 management:
   endpoints:
