+++ yorozuya-samples/micrometer-observation/src/test/resources/application.yml
@@ -8,20 +8,20 @@ spring:
       client:
         config:
           default:
-            connectTimeout: ${common.ok.http.client.connect-timeout}
-            read-timeout: ${common.ok.http.client.read-timeout}
+            connectTimeout: ${common.http.client.hc5.connect-timeout}
+            read-timeout: ${common.http.client.hc5.socket-timeout}
           feignClient:
             url: http://127.0.0.1:38080/rest
-      okhttp:
-        enabled: true
+      httpclient:
+        hc5:
+          enabled: true
 
 common:
-  ok:
-    http:
-      client:
+  http:
+    client:
+      hc5:
         connect-timeout: 3000
-        read-timeout: 5000
-        write-timeout: 5000
+        socket-timeout: 5000
   context:
     service:
       default-inject-key-set:
