--- yorozuya-samples/rest-client-httpclient/src/main/resources/application.yml
@@ -1,28 +1,0 @@
-spring:
-  application:
-    name: rest-client-httpclient
-  cloud:
-    openfeign:
-      client:
-        config:
-          #          default:
-          #            connectTimeout: ${common.http.client.connect-timeout}
-          #            read-timeout: ${common.http.client.socket-timeout}
-          #            logger-level: FULL
-          restClient:
-            url: http://127.0.0.1:8080/rest
-      httpclient:
-        hc5:
-          enabled: true
-
-common:
-  http:
-    client:
-      hc5:
-        enabled: true
-        connect-timeout: 3000
-        socket-timeout: 5000
-logging:
-  level:
-    feign: debug
-debug: true
