--- yorozuya-samples/rest-client-httpclient/src/test/java/com/github/ksewen/yorozuya/sample/rest/client/httpclient/RestClientHttpClientApplicationTests.java
@@ -1,7 +1,0 @@
-package com.github.ksewen.yorozuya.sample.rest.client.httpclient;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@SuppressWarnings("unused")
-class RestClientHttpClientApplicationTests {}
