--- yorozuya-samples/rest-client-okhttp/src/test/java/com/github/ksewen/yorozuya/sample/rest/client/okhttp/RestClientOkhttpApplicationTests.java
@@ -1,7 +1,0 @@
-package com.github.ksewen.yorozuya.sample.rest.client.okhttp;
-
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-@SuppressWarnings("unused")
-class RestClientOkhttpApplicationTests {}
