--- yorozuya-samples/rest-client-okhttp/src/test/resources/application.yml
@@ -1,24 +1,0 @@
-server:
-  port: 38080
-spring:
-  application:
-    name: rest-client-okhttp
-  cloud:
-    openfeign:
-      client:
-        config:
-          default:
-            connectTimeout: ${common.ok.http.client.connect-timeout}
-            read-timeout: ${common.ok.http.client.read-timeout}
-          feignClient:
-            url: http://127.0.0.1:38080/rest
-      okhttp:
-        enabled: true
-
-common:
-  ok:
-    http:
-      client:
-        connect-timeout: 3000
-        read-timeout: 5000
-        write-timeout: 5000
