+++ yorozuya-samples/rest-client/src/main/java/com/github/ksewen/yorozuya/sample/rest/client/httpclient/remote/ServerFacade.java
