+++ yorozuya-samples/rest-client/src/test/java/com/github/ksewen/yorozuya/sample/rest/client/httpclient/RestClientApplicationTests.java
@@ -1,0 +1,7 @@
+package com.github.ksewen.yorozuya.sample.rest.client.httpclient;
+
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+@SuppressWarnings("unused")
+class RestClientApplicationTests {}
