+++ yorozuya-samples/spring-data-jpa/src/main/resources/application.yml
@@ -2,4 +2,4 @@ spring:
   application:
     name: spring-data-jpa
 
-debug: true
+#debug: true
