+++ yorozuya-spring-boot-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -25,11 +25,8 @@ com.github.ksewen.yorozuya.starter.configuration.redis.RedisHelpersAutoConfigura
 
 # rest client
 com.github.ksewen.yorozuya.starter.configuration.http.client.HttpClientAutoConfiguration
-com.github.ksewen.yorozuya.starter.configuration.http.client.OkHttp3ClientAutoConfiguration
 com.github.ksewen.yorozuya.starter.configuration.http.client.RestTemplateAutoConfiguration
-com.github.ksewen.yorozuya.starter.configuration.http.client.RestTemplateAutoConfiguration.OkHttp3ClientRestTemplateCustomizerAutoConfiguration
 com.github.ksewen.yorozuya.starter.configuration.http.client.RestTemplateAutoConfiguration.HttpClientRestTemplateCustomizerAutoConfiguration
 
 # feign
-com.github.ksewen.yorozuya.starter.configuration.http.feign.OkHttp3FeignAutoConfiguration
 com.github.ksewen.yorozuya.starter.configuration.http.feign.HttpClient5FeignAutoConfiguration
