+++ yorozuya-spring-boot-starter/src/test/java/com/github/ksewen/yorozuya/starter/configuration/http/client/HttpClientRestTemplateCustomizerAutoConfigurationTest.java
@@ -28,8 +28,6 @@ class HttpClientRestTemplateCustomizerAutoConfigurationTest {
               assertThat(context)
                   .getBean("httpClientRestTemplateCustomizer")
                   .isSameAs(context.getBean(RestTemplateCustomizer.class));
-
-              assertThat(context).doesNotHaveBean("okHttp3ClientRestTemplateCustomizer");
             });
   }
 }
